feat: implement BLE device scanning functionality and add bleScan module
